Рефакторинг кода (из MVC в MVP): - Был создан MovieQuizPresenter - Метод convert(model: QuizQuestion) был перенесен в MovieQuizPresenter - Переменные questionsAmount и currentQuestionIndex также были перенесены в MovieQuizPresenter - Добавлено свойство presenter в MovieQuizViewControler - Применен DRYдля устранения дублирования данных.
